color, curves, rng, hash: edge cases
- cfromHex stopped reading at 8 digits, so "#RRGGBBAAjunk" parsed; the
  string must now end there. ctoColor8 cast NaN to uint8_t (undefined); NaN
  channels now encode as 0.
- ccubicEase falls back to bisection when Newton has not converged; where
  x(t) is flat it was off by ~1e-4 and is now within ~2e-6.
- ccatmullRom with a repeated end point (the usual way to start or end a
  spline) fell back to a straight lerp for the whole segment; the repeated
  point is now replaced by the reflected neighbour. chermite's comment had
  the tangents backwards.
- randInt computed hi - lo in int, overflowing for spans past INT_MAX.
- positionToCell divided by a zero bucketCount, and it and gridCellOf cast
  out-of-range or NaN floats to int; both clamp through the new cellCoordOf.
  hash.h includes what it uses.

+++ include/bromath/rng.h
@@ -35,8 +35,9 @@ inline float randRange(uint64_t& state, float lo, float hi) {
 
 inline int randInt(uint64_t& state, int loInclusive, int hiInclusive) {
     if (hiInclusive < loInclusive) return loInclusive;
-    uint64_t span = (uint64_t)(hiInclusive - loInclusive) + 1ULL;
-    return loInclusive + (int)(splitmix64(state) % span);
+    // In 64 bits: hi - lo overflows int for spans past INT_MAX.
+    uint64_t span = (uint64_t)((int64_t)hiInclusive - (int64_t)loInclusive) + 1ULL;
+    return (int)((int64_t)loInclusive + (int64_t)(splitmix64(state) % span));
 }
 
 // Box-Muller standard normal sample.
